+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Add `is_empty` to `Engine` API.
 * Add metadata deletion capability to `FileSystem` trait. Users can implement `exists_metadata` and `delete_metadata` to clean up obsolete metadata from older versions of Raft Engine.
+* Add `PerfContext` which records detailed time breakdown of the write process to thread-local storage.
 
 ## [0.2.0] - 2022-05-25